Handle invalid characters in path and file name

+++ utils/file.js
@@ -2,6 +2,14 @@ const fs = require('fs')
 const https = require('https')
 const path = require('path')
 
+const isDirVaild = (str) => {
+  const except = ['<', '>', ':', '/', '\\', '|', '?', '*', '"']
+  for (const i in except) {
+    if (str.includes(except[i])) { return false }
+  }
+  return true
+}
+
 const getParentDirectory = (filePath) => {
   const dir = path.dirname(filePath)
   console.log('dir:' + dir)
@@ -26,6 +34,7 @@ module.exports = {
       return JSON.parse(rawdata)
     } catch (error) {
       const dir = getParentDirectory(filePath)
+      if (!isDirVaild(dir)) { return {} }
       if (!fs.existsSync(dir)) { fs.mkdirSync(dir) }
       return {}
     }
@@ -37,27 +46,35 @@ module.exports = {
     } catch (error) {
       console.error(error)
       const dir = getParentDirectory(filePath)
+      if (!isDirVaild(dir)) { return }
       if (!fs.existsSync(dir)) { fs.mkdirSync(dir) }
       const data = JSON.stringify(obj)
       fs.writeFileSync(filePath, data)
     }
   },
   downloadFile: (url, dest, cb) => {
-    if (!checkURL(url)) { return }
-    if (!fs.existsSync(dest)) { fs.mkdirSync(dest) }
-    const fileDest = dest + uuidv4() + url.match(/\.(jpeg|jpg|gif|png)$/)[0]
-    var file = fs.createWriteStream(fileDest)
-    https.get(url, function (response) {
-      response.pipe(file)
-      file.on('finish', function () {
-        file.close(cb)
+    try {
+      if (!checkURL(url)) { return }
+      if (!isDirVaild(dest)) { return }
+      dest = 'assets/images/' + dest + '/'
+      if (!fs.existsSync(dest)) { fs.mkdirSync(dest) }
+      const fileDest = dest + uuidv4() + url.match(/\.(jpeg|jpg|gif|png)$/)[0]
+      var file = fs.createWriteStream(fileDest)
+      https.get(url, function (response) {
+        response.pipe(file)
+        file.on('finish', function () {
+          file.close(cb)
+        })
+      }).on('error', function (err) {
+        fs.unlink(fileDest)
+        if (cb) cb(err.message)
       })
-    }).on('error', function (err) {
-      fs.unlink(fileDest)
-      if (cb) cb(err.message)
-    })
+    } catch (error) {
+      console.log(error)
+    }
   },
   getRandomFile: (dir) => {
+    dir = 'assets/images/' + dir + '/'
     var files = fs.readdirSync(dir)
     const file = files[Math.floor(Math.random() * files.length)]
     if (file === undefined) { return null }
